Move classes to support package
