--- app/src/main/java/net/squanchy/extensions/Booleans.kt
@@ -1,7 +1,0 @@
-package net.squanchy.extensions
-
-inline fun Boolean.whenTrue(block: () -> Unit) {
-    if (this) {
-        block()
-    }
-}
