+++ app/src/main/java/net/squanchy/support/font/TypefaceDelegate.kt
@@ -1,8 +1,7 @@
-package net.squanchy.typeface
+package net.squanchy.support.font
 
 import android.graphics.Paint
 import android.graphics.Typeface
-import net.squanchy.extensions.whenTrue
 
 class TypefaceDelegate(private val newTypeface: Typeface) {
 
@@ -14,8 +13,13 @@ class TypefaceDelegate(private val newTypeface: Typeface) {
 
     fun applyTypefaceTo(paint: Paint) {
         val fakeStyle = fakeStyle(paint)
-        bold(fakeStyle).whenTrue { paint.isFakeBoldText = true }
-        italic(fakeStyle).whenTrue { paint.textSkewX = TEXT_SKEW_X }
+
+        if (bold(fakeStyle)) {
+            paint.isFakeBoldText = true
+        }
+        if (italic(fakeStyle)) {
+            paint.textSkewX = TEXT_SKEW_X
+        }
 
         paint.typeface = newTypeface
     }
@@ -24,8 +28,8 @@ class TypefaceDelegate(private val newTypeface: Typeface) {
 
     private fun italic(fakeStyle: Int) = fakeStyle.and(Typeface.ITALIC) != FALSE_FLAG
 
-    private fun fakeStyle(paint: Paint) : Int {
-        val oldStyle = paint.typeface?.style ?:0
+    private fun fakeStyle(paint: Paint): Int {
+        val oldStyle = paint.typeface?.style ?: 0
         return oldStyle.and(newTypeface.style.inv())
     }
 }
